Refactor: fix folder structure Refactor: error message refactoring
