+++ src/main/kotlin/com/esop/constant/Constants.kt
@@ -14,7 +14,11 @@ val errors = mapOf(
     "NO_ORDERS" to "User does not have any orders",
     "INVALID_TYPE" to "Type of Esop doesn't exist",
     "QUANTITY_NOT_ACCEPTED" to "Quantity of ESOPs must be between 0 to 10000",
-    "INVALID_USERNAME" to "Invalid User Name: username can consist only of alphabets, numbers or hyphen(s) and should start with an alphabet."
+    "INVALID_USERNAME" to "Invalid User Name: username can consist only of alphabets, numbers or hyphen(s) and should start with an alphabet.",
+    "WALLET_LIMIT_EXCEEDED" to "Wallet Limit exceeded",
+    "INVENTORY_LIMIT_EXCEEDED" to "Inventory Limit exceeded",
+    "POSITIVE_PLATFORM_FEE" to "Platform fee cannot be less than zero",
+    "INSUFFICIENT_FUNDS" to "Insufficient Funds"
 )
