+++ src/main/kotlin/com/esop/controller/UserController.kt
@@ -1,13 +1,11 @@
 package com.esop.controller
 
 
-import com.esop.HttpException
-import com.esop.InvalidPreOrderPlaceException
-import com.esop.UserDoesNotExistException
 import com.esop.dto.AddInventoryDTO
 import com.esop.dto.AddWalletDTO
 import com.esop.dto.CreateOrderDTO
 import com.esop.dto.UserCreationDTO
+import com.esop.exceptions.*
 import com.esop.schema.ESOPType
 import com.esop.schema.Order
 import com.esop.service.*
@@ -32,58 +30,6 @@ class UserController {
 
     @Inject
     lateinit var orderService: OrderService
-
-    @Error(exception = HttpException::class)
-    fun onHttpException(exception: HttpException): HttpResponse<Map<String, ArrayList<String?>>>? {
-        return HttpResponse.status<Map<String, ArrayList<String>>>(exception.status)
-            .body(mapOf("errors" to arrayListOf(exception.message)))
-    }
-
-    @Error(exception = JsonProcessingException::class)
-    fun onJSONProcessingException(ex: JsonProcessingException): HttpResponse<Map<String, ArrayList<String>>> {
-        return HttpResponse.badRequest(mapOf("errors" to arrayListOf("Invalid JSON format")))
-    }
-
-    @Error(exception = UnsatisfiedBodyRouteException::class)
-    fun onUnsatisfiedBodyRouteException(
-        request: HttpRequest<*>,
-        ex: UnsatisfiedBodyRouteException
-    ): HttpResponse<Map<String, List<String?>>> {
-        return HttpResponse.badRequest(mapOf("errors" to arrayListOf("Request body missing")))
-    }
-
-    @Error(status = HttpStatus.NOT_FOUND, global = true)
-    fun onRouteNotFoundException(): HttpResponse<Map<String, List<String?>>> {
-        return HttpResponse.notFound(mapOf("errors" to arrayListOf("Route not found")))
-    }
-
-    @Error(exception = ConversionErrorException::class)
-    fun onConversionErrorException(ex: ConversionErrorException): HttpResponse<Map<String, ArrayList<String?>>>? {
-        return HttpResponse.badRequest(mapOf("errors" to arrayListOf(ex.message)))
-    }
-
-    @Error(exception = ConstraintViolationException::class)
-    fun onConstraintViolationException(ex: ConstraintViolationException): HttpResponse<Map<String, List<String>>> {
-        return HttpResponse.badRequest(mapOf("errors" to ex.constraintViolations.map { it.message }))
-    }
-
-    @Error(exception = RuntimeException::class)
-    fun onRuntimeException(ex: RuntimeException): HttpResponse<Map<String, ArrayList<String?>>>? {
-        return HttpResponse.serverError(mapOf("errors" to arrayListOf(ex.message)))
-    }
-
-    @Error()
-    fun userDoesNotExistException(exception: UserDoesNotExistException): HttpResponse<Map<String, List<String>>>? {
-        return HttpResponse.badRequest(mapOf("error" to listOf(exception.errorMessage)))
-    }
-    @Error()
-    fun invalidPreOrderPlaceException(exception: InvalidPreOrderPlaceException): HttpResponse<Map<String, List<String>>>? {
-        return HttpResponse.badRequest(mapOf("error" to exception.errorList))
-    }
-    @Error(global = true)
-    fun globalError(e: Throwable): Map<String,List<String?>> {
-        return mapOf("error" to listOf(e.message))
-    }
 
 
     @Post(uri = "/register", consumes = [MediaType.APPLICATION_JSON], produces = [MediaType.APPLICATION_JSON])
