+++ src/test/kotlin/com/esop/InventoryTest.kt
@@ -1,6 +1,7 @@
 package com.esop
 
 import com.esop.constant.MAX_INVENTORY_CAPACITY
+import com.esop.exceptions.InventoryLimitExceededException
 import com.esop.schema.Inventory
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest
 import org.junit.jupiter.api.Assertions
