+++ src/test/kotlin/com/esop/schema/WalletTest.kt
@@ -1,13 +1,12 @@
 package com.esop.schema
 
-import com.esop.InsufficientFundsException
+import com.esop.exceptions.InsufficientFundsException
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
-import org.junit.jupiter.params.provider.ValueSource
 import java.util.stream.Stream
 
 class WalletTest {
