+++ src/test/kotlin/com/esop/service/UserServiceTest.kt
@@ -1,11 +1,11 @@
 package com.esop.service
 
-import com.esop.*
 import com.esop.constant.MAX_INVENTORY_CAPACITY
 import com.esop.constant.MAX_WALLET_CAPACITY
 import com.esop.dto.AddInventoryDTO
 import com.esop.dto.AddWalletDTO
 import com.esop.dto.UserCreationDTO
+import com.esop.exceptions.*
 import com.esop.repository.UserRecords
 import com.esop.schema.Order
 import org.junit.jupiter.api.Assertions.assertEquals
